Added install script and .bash_profile.
Added .bash_profile for non interactive shell.

The install script first do a preflight test to see if any files or directories already exists.
Then, the bin/ and dotfiles/ directories are copied to the current user's home directory. And symlinks are created for .bashrc and .bash_profile to their homonymes in dotfiles/.
Also began to implement option for remote install.