Added information to download the dotfiles as a tarball. Install script updated to work with all download methods...
Not everyone or every host has git installed. And it maybe overkill
to install Git only to manage the dotifles. So i added a tarball
download method in the readme.

But then, this may lead to a dangerous move and rename operation who
could overwrite or discard existing local files. The install script
has been updated to deal with this. Not by doing a proper update, but
by providing a way to safely backup old files.

- Force install now backup ~/.dotfiles as well, but only if install is run from another directory
- Backup directory name is now tagged with a timestamp
- Create ~/.dotfiles and copy content in it if not existing
- Added note about changing prompt color